Move each piece to a trait
