+++ src/lib/Client/DaprHttpClient.php
@@ -3,11 +3,8 @@
 namespace Dapr\Client;
 
 use Dapr\Deserialization\IDeserializer;
-use Dapr\exceptions\DaprException;
 use Dapr\Serialization\ISerializer;
 use GuzzleHttp\Client;
-use GuzzleHttp\Promise\PromiseInterface;
-use Psr\Http\Message\ResponseInterface;
 
 /**
  * Class DaprHttpClient
@@ -18,6 +15,8 @@ class DaprHttpClient extends DaprClient
     use HttpStateTrait;
     use HttpSecretsTrait;
     use HttpInvokeTrait;
+    use HttpPubSubTrait;
+    use HttpBindingTrait;
 
     private Client $httpClient;
 
@@ -37,40 +36,6 @@ class DaprHttpClient extends DaprClient
                     'Content-Type' => 'application/json'
                 ]
             ]
-        );
-    }
-
-    /**
-     * @throws DaprException
-     */
-    public function invokeMethod(
-        string $httpMethod,
-        string $appId,
-        string $methodName,
-        mixed $data = null,
-        array $metadata = []
-    ): ResponseInterface {
-        return $this->handlePromise(
-            $this->invokeMethodAsync($httpMethod, $appId, $methodName, $data, $metadata)
-        )->wait();
-    }
-
-    public function invokeMethodAsync(
-        string $httpMethod,
-        string $appId,
-        string $methodName,
-        mixed $data = null,
-        array $metadata = []
-    ): PromiseInterface {
-        $options = [];
-        if (!empty($data)) {
-            $options['body'] = $this->serializer->as_json($data);
-        }
-        $options['headers'] = $metadata;
-        $appId = rawurlencode($appId);
-        $methodName = rawurlencode($methodName);
-        return $this->handlePromise(
-            $this->httpClient->requestAsync($httpMethod, "/v1.0/invoke/$appId/method/$methodName", $options)
         );
     }
 }
