+++ src/lib/Client/HttpInvokeTrait.php
@@ -2,9 +2,6 @@
 
 namespace Dapr\Client;
 
-use Dapr\Client\BindingRequest;
-use Dapr\Client\BindingResponse;
-use Dapr\Client\PromiseHandlingTrait;
 use Dapr\Deserialization\IDeserializer;
 use Dapr\exceptions\DaprException;
 use Dapr\Serialization\ISerializer;
@@ -16,7 +13,8 @@ use Psr\Http\Message\ResponseInterface;
  * Trait HttpInvokeTrait
  * @package Dapr\Client
  */
-trait HttpInvokeTrait {
+trait HttpInvokeTrait
+{
     use PromiseHandlingTrait;
 
     public IDeserializer $deserializer;
@@ -26,54 +24,34 @@ trait HttpInvokeTrait {
     /**
      * @throws DaprException
      */
-    public function invokeBinding(BindingRequest $bindingRequest, string $dataType = 'array'): BindingResponse
-    {
-        return $this->invokeBindingAsync($bindingRequest, $dataType)->wait();
-    }
-
-    public function invokeBindingAsync(BindingRequest $bindingRequest, string $dataType = 'array'): PromiseInterface
-    {
+    public function invokeMethod(
+        string $httpMethod,
+        string $appId,
+        string $methodName,
+        mixed $data = null,
+        array $metadata = []
+    ): ResponseInterface {
         return $this->handlePromise(
-            $this->httpClient->putAsync(
-                '/v1.0/bindings/' . rawurlencode($bindingRequest->bindingName),
-                [
-                    'body' => $this->serializer->as_json(
-                        [
-                            'data' => $bindingRequest->data,
-                            'metadata' => $bindingRequest->metadata,
-                            'operation' => $bindingRequest->operation
-                        ]
-                    )
-                ]
-            ),
-            fn(ResponseInterface $response) => new BindingResponse(
-                $bindingRequest,
-                $this->deserializer->from_json($dataType, $response->getBody()->getContents()),
-                $response->getHeaders()
-            )
-        );
+            $this->invokeMethodAsync($httpMethod, $appId, $methodName, $data, $metadata)
+        )->wait();
     }
 
-    /**
-     * @throws DaprException
-     */
-    public function publishEvent(string $pubsubName, string $topicName, mixed $data, array $metadata = []): void
-    {
-        $this->publishEventAsync($pubsubName, $topicName, $data, $metadata)->wait();
-    }
-
-    public function publishEventAsync(
-        string $pubsubName,
-        string $topicName,
-        mixed $data,
+    public function invokeMethodAsync(
+        string $httpMethod,
+        string $appId,
+        string $methodName,
+        mixed $data = null,
         array $metadata = []
     ): PromiseInterface {
-        $options = [
-            'query' => $metadata,
-            'body' => $this->serializer->as_json($data)
-        ];
-        $pubsubName = rawurlencode($pubsubName);
-        $topicName = rawurlencode($topicName);
-        return $this->handlePromise($this->httpClient->postAsync("/v1.0/publish/$pubsubName/$topicName", $options));
+        $options = [];
+        if (!empty($data)) {
+            $options['body'] = $this->serializer->as_json($data);
+        }
+        $options['headers'] = $metadata;
+        $appId = rawurlencode($appId);
+        $methodName = rawurlencode($methodName);
+        return $this->handlePromise(
+            $this->httpClient->requestAsync($httpMethod, "/v1.0/invoke/$appId/method/$methodName", $options)
+        );
     }
 }
